Correctly add `GIT_COMMIT_HASH` into version string:
Commit bf013c02ad8e49b9242d73afae245e48fa0ad5b3 added support
for incorporating a commit ID into the compiled version string
but did so in a way that did not follow the semantic versioning
standard.

This commit corrects that flaw by moving the commit ID into the
"metadata" part of the version string and properly handles the
case where the commit hash cannot be retrieved.